+++ socketscpi/__init__.py
@@ -1,6 +1,6 @@
 __author__ = """Morgan Allison"""
 __email__ = 'morgan.allison@keysight.com'
-__version__ = '2023.06.0'
+__version__ = '2024.11.0'
 
 # For ease of use
 from socketscpi.socketscpi import *